Separate ScoresList into modules, loose localStorage functionality.
